% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{compare_pred_sets}
\alias{compare_pred_sets}
\title{Title}
\usage{
compare_pred_sets(density_p0 = list(type = "normal", mean = 0, sd = 1),
  density_q = list(type = "normal", mean = 5, sd = 1), eps = 0.2,
  alpha = 0.05, tol = 0.001, scale_factor = 5)
}
\arguments{
\item{scale_factor}{}
}
\description{
Title
}
