% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{gen_density}
\alias{gen_density}
\title{Generate the required density function}
\usage{
gen_density(density_list)
}
\arguments{
\item{density_list}{}
}
\description{
Generate the required density function
}
