% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-huber.R
\name{gen_density_samples}
\alias{gen_density_samples}
\title{Generate specified number of samples from the required density function}
\usage{
gen_density_samples(density_list, n_samples = 100)
}
\arguments{
\item{n_samples}{}
}
\description{
Generate specified number of samples from the required density function
}
