% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{gen_distribution}
\alias{gen_distribution}
\title{Generate the specified distribution}
\usage{
gen_distribution(density_list)
}
\arguments{
\item{density_list}{}
}
\description{
Generate the specified distribution
}
