% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{generate_huber_density}
\alias{generate_huber_density}
\title{Generate the required huber density function}
\usage{
generate_huber_density(density_p0 = list(type = "normal", mean = 0, sd =
  1), density_q = list(type = "normal", mean = 0, sd = 1), eps)
}
\arguments{
\item{eps}{}
}
\description{
Generate the required huber density function
}
