% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conformal.R
\name{get_conf_approx}
\alias{get_conf_approx}
\title{Performs Approximate Conformal Prediction}
\usage{
get_conf_approx(input_data, h, alpha, grid_xmin, grid_xmax, grid_bgridsize)
}
\arguments{
\item{input_data}{d by n data matrix}

\item{h}{bandwidth for Kernel Density Estimation}

\item{alpha}{prediction significance level}

\item{grid}{d by n_grid coordinate grids}
}
\description{
Performs Approximate Conformal Prediction
}
