% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{get_density_plot_support}
\alias{get_density_plot_support}
\title{Title}
\usage{
get_density_plot_support(density_list, scale_factor = 5)
}
\arguments{
\item{scale_factor}{}
}
\description{
Title
}
