% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{get_huber_plot_support}
\alias{get_huber_plot_support}
\title{Title}
\usage{
get_huber_plot_support(density_p0 = list(type = "normal", mean = 0, sd =
  1), density_q = list(type = "normal", mean = 5, sd = 1),
  scale_factor = 5)
}
\arguments{
\item{scale_factor}{}
}
\description{
Title
}
