% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{get_known_level_set}
\alias{get_known_level_set}
\title{Get the level set for the given density type}
\usage{
get_known_level_set(density_list, alpha)
}
\arguments{
\item{alpha}{}
}
\description{
Get the level set for the given density type
}
