% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{get_level_set}
\alias{get_level_set}
\title{Generate the level set for the Huber density}
\usage{
get_level_set(huber_density, level, tol, grid_params)
}
\arguments{
\item{grid_params}{}
}
\description{
Generate the level set for the Huber density
}
