% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{get_level_set_probability_mass}
\alias{get_level_set_probability_mass}
\title{Get probability mass for the level set}
\usage{
get_level_set_probability_mass(level_set, huber_distribution)
}
\arguments{
\item{huber_distribution}{}
}
\description{
Get probability mass for the level set
}
