% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uci.R
\name{get_plot_clss}
\alias{get_plot_clss}
\title{Create the scatter plot of the True and \code{epsilon}-contaminated
binary class dataset split by the specified 2 variables}
\usage{
get_plot_clss(ds, ds_name, plt_vars, clss_vals, clss_rnm_vals, epsilon,
  tot_nrows_req)
}
\arguments{
\item{ds}{(data.frame) : The input cleaned data frame, assumes that this has
a binary valued variable called \code{class}}

\item{ds_name}{(character) : The name of the data set we are importing to be
displayed in the plot title}

\item{plt_vars}{(character) : The 2 variables we intend to plot across the
class labels in our final dataset, specified in the x-y ordering}

\item{clss_vals}{(character) : These are the binary class variables values.
The first element value is assumed to be the true class value and
the second value is assumed to be the contamination class value}

\item{clss_rnm_vals}{(character) : These are the renamed binary class
variables values to be displayed in the plot legend. The first element value
is assumed to be the true class value and the second value is assumed to be
the contamination class value.}

\item{epsilon}{(double) : The contamination proportion, must be a value in
(0, 1), but should really be < 0.5}

\item{tot_nrows_req}{(integer) : The total number of observations i.e. True
and contamination values to display in the final dataset and plot}
}
\description{
Create the scatter plot of the True and \code{epsilon}-contaminated
binary class dataset split by the specified 2 variables
}
