% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-robust-pred-sets.R
\name{get_robust_prediction_set}
\alias{get_robust_prediction_set}
\title{Get the robust prediction set}
\usage{
get_robust_prediction_set(density_p0, density_q, eps, alpha, grid_params,
  tol = 0.001)
}
\arguments{
\item{tol}{}
}
\description{
Get the robust prediction set
}
