% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uci.R
\name{get_split_clss}
\alias{get_split_clss}
\title{Get a list of sampled dataframes of the original binary class dataframe
split to get the specified true and contamination proportion between the
2 binary classes.}
\usage{
get_split_clss(uci_dat, plt_vars, clss_vals, epsilon, tot_nrows_req)
}
\arguments{
\item{uci_dat}{(dataframe) : Dataframe for the binary classification data}

\item{plt_vars}{(character) : These are the 2 variables we will scatterplot
User needs to ensure that the class variable is last in the vector i.e. the
third value}

\item{clss_vals}{(character) : These are the binary class variables values.
The first element value is assumed to be the true class value and
the second value is assumed to be the contamination class value}

\item{epsilon}{(double) : The proportion of the contamination class value that
we require in our final dataframe}

\item{tot_nrows_req}{(integer) : The total number of observations we require
in our final dataframe of true and contaminated proportions}
}
\value{
(dataframe) : Combined resampled dataframe of true and contamination
classes
}
\description{
Get a list of sampled dataframes of the original binary class dataframe
split to get the specified true and contamination proportion between the
2 binary classes.
}
