% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uci.R
\name{get_uci_data}
\alias{get_uci_data}
\title{Function to download and clean UCI data into a tibble}
\usage{
get_uci_data(ucidurl, col_names_raw, col_types_raw)
}
\arguments{
\item{ucidurl}{(character): the specific url subdirectory of the required
UCI dataset e.g. for the UCI hepatitis dataset, is typically located in the
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/hepatitis/hepatitis.data}
so user only needs to enter "hepatitis/hepatitis.data" here in this case}

\item{col_names_raw}{(character): A vector of column names for the input
dataset. These can be in any case. Spaces should be replaced by uncerscores
for all values e.g. c("Body Temp") should be changed to c("Body_Temp")}

\item{col_types_raw}{(character): A single string where each character
corresponds to the \code{readr} shorthand for the column type
(e.g. "f" for factor)#' for the corresponding column to be read in by
\code{readr}}
}
\value{
tibble: required UCI dataset
}
\description{
Function to download and clean UCI data into a tibble
}
