% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uci.R
\name{split_by_class}
\alias{split_by_class}
\title{Returns an original binary class dataframe filtered to the required
\code{class} variable column}
\usage{
split_by_class(df, clss_val)
}
\arguments{
\item{df}{(dataframe) : The required binary class dataframe to be filtered.
This is required to have a \code{class} variable}

\item{clss_val}{(character) : The required class value to be filtered}
}
\value{
(dataframe) : The filtered version of the original dataframe to the
required class value
}
\description{
Returns an original binary class dataframe filtered to the required
\code{class} variable column
}
